Fix Issues with layout on Android View.
